+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 * Add the `unified_admin_domain` configuration option for the unified admin domain.
+- Add new generators for webhook subscriptions defined in the `shopify.app.toml` file [1882](https://github.com/Shopify/shopify_app/pull/1882)
 
 22.3.1 (July 26, 2024)
 ----------